refactor(meta): splits source_manager into smaller mods
